# openapi.yaml
openapi: 3.0.0
info:
  x-skaffolder-id-project: 6016ab4ba3e049112323e4ce
  title: CMO
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  /user:
    post:
      x-skaffolder-id: 6016ab4ba3e049112323e4dd
      x-skaffolder-name: create
      x-skaffolder-id-resource: 6016ab4ba3e049112323e4d8
      x-skaffolder-resource: User
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - User
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                mail:
                  type: string
                name:
                  type: string
                password:
                  type: string
                roles:
                  type: string
                surname:
                  type: string
                username:
                  type: string
    get:
      x-skaffolder-id: 6016ab4ba3e049112323e4e0
      x-skaffolder-name: list
      x-skaffolder-id-resource: 6016ab4ba3e049112323e4d8
      x-skaffolder-resource: User
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - User
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  "/user/{id}":
    delete:
      x-skaffolder-id: 6016ab4ba3e049112323e4dc
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 6016ab4ba3e049112323e4d8
      x-skaffolder-resource: User
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 6016ab4ba3e049112323e4de
      x-skaffolder-name: get
      x-skaffolder-id-resource: 6016ab4ba3e049112323e4d8
      x-skaffolder-resource: User
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    post:
      x-skaffolder-id: 6016ab4ba3e049112323e4df
      x-skaffolder-name: update
      x-skaffolder-id-resource: 6016ab4ba3e049112323e4d8
      x-skaffolder-resource: User
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      x-skaffolder-id: 6016ab4ba3e049112323e4d8
      x-skaffolder-id-db: 6016ab4ba3e049112323e4cf
      x-skaffolder-id-entity: 6016ab4ba3e049112323e4d0
      x-skaffolder-url: /user
      x-skaffolder-type: User
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        mail:
          type: string
          x-skaffolder-id-attr: 6016ab4ba3e049112323e4d3
          x-skaffolder-type: String
        name:
          type: string
          x-skaffolder-id-attr: 6016ab4ba3e049112323e4d4
          x-skaffolder-type: String
        password:
          type: string
          x-skaffolder-id-attr: 6016ab4ba3e049112323e4d2
          x-skaffolder-type: String
          x-skaffolder-required: true
        roles:
          type: string
          x-skaffolder-id-attr: 6016ab4ba3e049112323e4d6
          x-skaffolder-type: String
        surname:
          type: string
          x-skaffolder-id-attr: 6016ab4ba3e049112323e4d5
          x-skaffolder-type: String
        username:
          type: string
          x-skaffolder-id-attr: 6016ab4ba3e049112323e4d1
          x-skaffolder-type: String
          x-skaffolder-required: true
      x-skaffolder-relations: 
      required:
        - _id
        - password
        - username
...